Status LED color is now correct
Changed the input resistance value on the breadboard to limit the power
to the yellow RGB pin. Added a lockKeypad function to keypadTest.h that
locks the keypad until the correct pin is entered again. Added a switch
case in main so that the keypad locks if 'D' is pressed.