Added different button functionality to keypad
Pressing D resets the system, pressing A turns on an LED on the LED bar.
I meant to be working in a different branch, I didnt realize I was in
main. Added a turn on LED and turn off LED function to LEDbar.h that
intakes an int and changes the corresponding LED. Test_project/ is no
longer recognized by CCS, so all my changes had to be done in app/